panel

+++ app/views/categories/index.blade.php
@@ -1,13 +1,16 @@
 @section('content')
-
-<a href="{{ route('categories.create') }}" style="color: #666; margin-left: 10px;">NEW CATEGORY</a>
+<div class="tooltip-div">
+	<a href="{{ route('categories.create') }}" class="btn pull-right" data-placement="left" title="add category"><span class="glyphicon glyphicon-plus"></span></a>
+</div>
+<div style="clear:both"></div>
 
 @if ($categories->count())
 	<table class="table">
 		<thead>
 			<tr>
-				<th>Name</th>
-				<th>Status</th>
+				<th>name</th>
+				<th>articles</th>
+				<th>deleted</th>
 				<th></th>
 				<th></th>
 			</tr>
@@ -17,13 +20,14 @@
 			@foreach ($categories as $category)
 				<tr>
 					<td>{{{ $category->name }}}</td>
-					<td>{{{ $category->status }}}</td>
-                    <td><a href="{{ route('categories.edit', $category->id) }}" class="btn btn-info">edit</a></td>
-                    <td>
-                        {{ Form::open(['method' => 'DELETE', 'route' => ['categories.destroy', $category->id]]) }}
-                            {{ Form::submit(trans('messages.delete'), ['class' => 'btn btn-danger']) }}
-                        {{ Form::close() }}
-                    </td>
+					<td>{{{ $category->articles->count() }}}</td>
+					<td>{{{ $category->deleted_at }}}</td>
+					<td><a href="{{ route('categories.edit', $category->id) }}" class="btn btn-primary">edit</a></td>
+					<td>
+						{{ Form::open(['method' => 'DELETE', 'route' => ['categories.destroy', $category->id], 'onsubmit' => 'return confirm("Are you sure?")']) }}
+							<button type="submit" class="btn btn-danger">delete</button>
+						{{ Form::close() }}
+					</td>
 				</tr>
 			@endforeach
 		</tbody>
@@ -32,4 +36,13 @@
 	There are no categories
 @endif
 
+@stop
+
+@section('script')
+<script type="text/javascript">
+$(".tooltip-div").tooltip({
+	selector: "button, a",
+	container: "body"
+});
+</script>
 @stop
